feat: update the `zisk` crypto backend for `ethrex`
